AIHold-Back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
pydantic==2.10.0
pydantic[email]==2.10.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
brapi==1.0.0
requests==2.31.0

